Move the IDocumentationCommentSnippetService down to the Features layer
This service does not depend on anything in the EditorFeatures layer, and can be easily ported over with just a couple of OptionsProvider fixups.
