+++ src/Features/Core/Portable/DocumentationComments/IDocumentationCommentSnippetService.cs
@@ -7,7 +7,7 @@ using Microsoft.CodeAnalysis.Host;
 using Microsoft.CodeAnalysis.Options;
 using Microsoft.CodeAnalysis.Text;
 
-namespace Microsoft.CodeAnalysis.Editor
+namespace Microsoft.CodeAnalysis.DocumentationComments
 {
     internal interface IDocumentationCommentSnippetService : ILanguageService
     {
