+++ src/VisualStudio/CSharp/Impl/Options/AutomationObject.cs
@@ -13,6 +13,7 @@ using Microsoft.CodeAnalysis.Completion;
 using Microsoft.CodeAnalysis.CSharp.CodeStyle;
 using Microsoft.CodeAnalysis.CSharp.Formatting;
 using Microsoft.CodeAnalysis.Diagnostics.Analyzers.NamingStyles;
+using Microsoft.CodeAnalysis.DocumentationComments;
 using Microsoft.CodeAnalysis.Editing;
 using Microsoft.CodeAnalysis.Editor.Shared.Options;
 using Microsoft.CodeAnalysis.ExtractMethod;
@@ -47,8 +48,8 @@ namespace Microsoft.VisualStudio.LanguageServices.CSharp.Options
 
         public int AutoComment
         {
-            get { return GetBooleanOption(FeatureOnOffOptions.AutoXmlDocCommentGeneration); }
-            set { SetBooleanOption(FeatureOnOffOptions.AutoXmlDocCommentGeneration, value); }
+            get { return GetBooleanOption(DocumentationCommentOptions.AutoXmlDocCommentGeneration); }
+            set { SetBooleanOption(DocumentationCommentOptions.AutoXmlDocCommentGeneration, value); }
         }
 
         public int AutoInsertAsteriskForNewLinesOfBlockComments
